update the type  of id from int to bigint and update user table

+++ blog-portal/src/main/java/site/iblogs/portal/controller/CommentController.java
@@ -27,7 +27,7 @@ public class CommentController {
     @ApiOperation("通过文章Id获取评论")
     @RequestMapping(value = "/getComments", method = RequestMethod.GET)
     @ResponseBody
-    public ApiResponse<PageResponse<CommentResponse>> GetComments(int cid, int pageNum, int pageSize){
+    public ApiResponse<PageResponse<CommentResponse>> GetComments(Long cid, int pageNum, int pageSize){
         return ApiResponse.success(commentService.getComment(cid,pageNum,pageSize));
     }
